+++ .eslintrc.js
@@ -15,6 +15,7 @@ module.exports = {
     '@typescript-eslint/no-unused-vars': 'off',
     '@typescript-eslint/no-unused-vars-experimental': 'off',
     '@typescript-eslint/no-misused-promises': 'off',
+    '@typescript-eslint/no-extraneous-class': 'off',
     '@typescript-eslint/explicit-function-return-type': 'off',
     '@typescript-eslint/prefer-reduce-type-parameter': 'off',
     '@typescript-eslint/strict-boolean-expressions': 'off',