+++ packages/mobile/codegen/PlopGeneratorConfig.type.ts
@@ -1,0 +1,6 @@
+import {PlopGenerator} from 'plop';
+
+export type PlopGeneratorConfig = Omit<
+  PlopGenerator,
+  'runActions' | 'runPrompts'
+>;