+++ packages/mobile/codegen/function/templates/template.function.ts
@@ -1,0 +1,5 @@
+import {functionConst} from '../function.const';
+
+export const template = `
+export const {{ ${functionConst.vars.name} }} = (): void => {}
+`;