+++ packages/mobile/metro.config.js
@@ -1,0 +1,27 @@
+/**
+ * Metro configuration for React Native
+ * https://github.com/facebook/react-native
+ *
+ * @format
+ */
+
+const blacklist = require('metro-config/src/defaults/exclusionList');
+
+module.exports = {
+  resolver: {
+    blacklistRE: blacklist([
+      /\/nodejs-assets\/.*/,
+      /\/android\/.*/,
+      /\/ios\/.*/,
+    ]),
+    sourceExts: ['jsx', 'js', 'ts', 'tsx'],
+  },
+  transformer: {
+    getTransformOptions: async () => ({
+      transform: {
+        experimentalImportSupport: false,
+        inlineRequires: true,
+      },
+    }),
+  },
+};