+++ packages/mobile/src/components/Message/Message.types.ts
@@ -1,0 +1,5 @@
+import { DisplayableMessage } from '@quiet/nectar'
+
+export interface MessageProps {
+  message: DisplayableMessage
+}