+++ packages/nectar/src/sagas/publicChannels/markUnreadMessages/unreadMessages.adapter.ts
@@ -1,0 +1,5 @@
+import { createEntityAdapter } from '@reduxjs/toolkit'
+import { UnreadChannelMessage } from '../publicChannels.types'
+
+export const unreadMessagesAdapter =
+    createEntityAdapter<UnreadChannelMessage>()